+++ lib/dart_extension_methods.dart
@@ -1,0 +1,15 @@
+library dart_extensions_methods;
+
+import 'dart:collection';
+import 'dart:core';
+export 'package:dart_extensions_methods/dart_extension_methods.dart';
+
+part 'iterable_methods.dart';
+
+part 'string_methods.dart';
+
+part 'general.dart';
+
+part 'boolean_methods.dart';
+
+part 'pair.dart';